requirements.txt
django
djangorestframework
pytz
python-telegram-bot
requests